pyproject.toml
[project]
dependencies = [
    "moviepy",
    "plotnine",
    "typer",
    "ipython",
    "ipykernel",
    "polars",
    "pyarrow",
    "PyYAML",
]
name = "vlm-benchmark"
version = "0.0.1"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
alfred = "main:app"
